% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{ab_reckoning}
\alias{ab_reckoning}
\title{accessor for reckoning element}
\usage{
ab_reckoning(x)
}
\arguments{
\item{x}{an instance of avReckoning}
}
\value{
a tibble with one row for each expense type by time slice
}
\description{
accessor for reckoning element
}
\examples{
dim(ab_reckoning(demo_rec))
}
