% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_reck}
\alias{browse_reck}
\title{prototypical cost exploring app}
\usage{
browse_reck()
}
\value{
returns "NULL"
}
\description{
prototypical cost exploring app
}
\examples{
if (interactive()) browse_reck()
}
