% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_rec}
\alias{demo_rec}
\title{a demonstration avReckoning object}
\format{avReckoning instance}
\usage{
demo_rec
}
\description{
a demonstration avReckoning object
}
\note{
This is a snapshot of cost data collected for a specific project.
}
\examples{
demo_rec
}
\keyword{datasets}
