% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billing.R
\name{getSkus}
\alias{getSkus}
\title{List the available GCP product skus}
\usage{
getSkus(mybilling)
}
\arguments{
\item{mybilling}{tbl_df}
}
\value{
character()
}
\description{
List the available GCP product skus
}
