% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billing.R
\name{getSubmissionCost}
\alias{getSubmissionCost}
\title{Calcuate costs for a workflow submission by ID}
\usage{
getSubmissionCost(mybilling, submissionID)
}
\arguments{
\item{mybilling}{tbl_df}

\item{submissionID}{character(1) Terra submission ID}
}
\value{
numeric()
}
\description{
Calcuate costs for a workflow submission by ID
}
\examples{
data(demo_rec) # makes rec
v = getValues(demo_rec@reckoning, "terra-submission-id")[1] # for instance
getSubmissionCost(demo_rec@reckoning,v)
}
