% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billing.R
\name{getSubmissionRam}
\alias{getSubmissionRam}
\title{Calcuate ram usage for a workflow submission by ID}
\usage{
getSubmissionRam(mybilling, submissionID)
}
\arguments{
\item{mybilling}{tbl_df}

\item{submissionID}{character(1) Terra submission ID}
}
\value{
data.frame
}
\description{
Calcuate ram usage for a workflow submission by ID
}
\examples{
data(demo_rec) # makes rec
v = getValues(demo_rec@reckoning, "terra-submission-id")[1] # for instance
getSubmissionRam(demo_rec@reckoning,v)
}
