% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billing.R
\name{getValues}
\alias{getValues}
\title{deal with nested tables in a reckoning}
\usage{
getValues(mybilling, mykey)
}
\arguments{
\item{mybilling}{tbl_df from reckon()}

\item{mykey}{character(1) key}
}
\value{
character()
}
\description{
deal with nested tables in a reckoning
}
\examples{
if (interactive()) #' getValues(reckoning(demo_rec), "security")
}
