% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{reckon}
\alias{reckon}
\title{perform reckoning}
\usage{
reckon(obj)
}
\arguments{
\item{obj}{instance of avReckoningRequest}
}
\description{
perform reckoning
}
\examples{
req1 = setup_billing_request( "2020-01-28", "2020-01-29",
 "bjbilling", "anvilbilling", "gcp_billing_export_v1_015E39_38569D_3CC771",
 "landmarkanvil2")
rec = reckon(req1)
}
