% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{reckoning,avReckoning-method}
\alias{reckoning,avReckoning-method}
\title{accessor for reckoning component}
\usage{
\S4method{reckoning}{avReckoning}(x)
}
\arguments{
\item{x}{instance of avReckoning}
}
\value{
tbl_df
}
\description{
accessor for reckoning component
}
\examples{
if (interactive()) reckoning(reckon(demo_rec))
}
