% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{reckoning}
\alias{reckoning}
\title{generic for accessor for reckoning component}
\usage{
reckoning(x)
}
\arguments{
\item{x}{object inheriting from avReckoning}
}
\value{
tbl_df
}
\description{
generic for accessor for reckoning component
}
\examples{
if (interactive()) reckoning(reckon(demo_rec))
}
