% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billing.R
\name{subsetBySku}
\alias{subsetBySku}
\title{subset a billing object by sku}
\usage{
subsetBySku(mybilling, mysku)
}
\arguments{
\item{mysku}{character(1) GCP product sku}

\item{mybillilng}{tbl_df}
}
\description{
subset a billing object by sku
}
