% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_reck}
\alias{browse_reck}
\title{prototypical cost exploring app}
\usage{
browse_reck(bq_email = NA)
}
\arguments{
\item{bq_email}{character(1) email used to identify google BigQuery api user}
}
\description{
prototypical cost exploring app
}
