% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billing.R
\name{getBilling}
\alias{getBilling}
\title{request billing data}
\usage{
getBilling(startDate, endDate, bqProject, bqDataset, bqTable, bqBilling)
}
\arguments{
\item{start}{character(1) date of start of reckoning}

\item{end}{character(1) date of end of reckoning}

\item{project}{character(1) GCP project id}

\item{dataset}{character(1) GCP dataset id for billing data in BQ}

\item{table}{character(1) GCP table for billing data in BQ}

\item{billing_code}{character(1) GCP billing code}
}
\description{
request billing data
}
