% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{setup_billing_request}
\alias{setup_billing_request}
\title{set up request object}
\usage{
setup_billing_request(start, end, project, dataset, table, billing_code)
}
\arguments{
\item{start}{character(1) date of start of reckoning}

\item{end}{character(1) date of end of reckoning}

\item{project}{character(1) GCP project id}

\item{dataset}{character(1) GCP dataset id for billing data in BQ}

\item{table}{character(1) GCP table for billing data in BQ}

\item{billing_code}{character(1) GCP billing code}
}
\description{
set up request object
}
\examples{
req1 = setup_billing_request( "2020-01-28", "2020-01-29",
 "bjbilling", "anvilbilling", "gcp_billing_export_v1_015E39_38569D_3CC771",
 "landmarkanvil2")
req1
}
